backend/requirements.txt
# FastAPI and ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Pydantic
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25

# Authentication
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.6

# Logging
structlog==24.1.0

# PDF Generation (preserved from original)
reportlab==4.3.1
pillow==11.1.0

# Data processing and exports
pandas==2.2.3
openpyxl==3.1.5
xlsxwriter==3.1.2

# Utilities
num2words==0.5.14
pytz==2025.1

# Testing
pytest==8.3.5
pytest-asyncio==0.25.0
pytest-cov==6.1.0
httpx==0.27.2
